fix: put bit operations to a separate header file
